Added PlayfieldSwitchEnum.h

+++ lib/Domain/SchedulerRegistry.cpp
@@ -1,5 +1,6 @@
 #include "SchedulerRegistry.h"
 #include <propeller.h>
+#include <stdint.h>
 
 SchedulerRegistry::SchedulerRegistry() {
   _schedulersInList = 0;
@@ -11,7 +12,7 @@ SchedulerRegistry::~SchedulerRegistry() {
 
 
 void SchedulerRegistry::clear() {
-  for (int i = 0; i <= MAX_SCHEDULERS; i++) {
+  for (uint8_t i = 0; i <= MAX_SCHEDULERS; i++) {
     _pSchedulers[i] = 0;
   }
   _schedulersInList = 0;
@@ -23,7 +24,7 @@ void SchedulerRegistry::addScheduler (Scheduler &scheduler, uint16_t updateInter
 }
 
 void SchedulerRegistry::schedule() {
-  for (int i = 0; i < _schedulersInList; i++) {
+  for (uint8_t i = 0; i < _schedulersInList; i++) {
     if (_pSchedulers[i]->shouldSchedule()) {
       _pSchedulers[i]->schedule();
     }
